separate to utils file
