+++ CMakeLists.txt
@@ -109,6 +109,8 @@ add_library(viam-ur)
 target_sources(viam-ur
   PRIVATE
     src/viam/ur/module/ur_arm.cpp
+    src/viam/ur/module/utils.hpp
+
 )
 target_link_libraries(viam-ur
   PUBLIC
