+++ src/viam/ur/module/ur_arm.cpp
@@ -13,19 +13,19 @@
 #include <boost/numeric/conversion/cast.hpp>
 #include <boost/range/adaptor/transformed.hpp>
 
-#include <ur_client_library/log.h>
 #include <ur_client_library/types.h>
 #include <ur_client_library/ur/dashboard_client.h>
 #include <ur_client_library/ur/ur_driver.h>
 
 #include <viam/sdk/components/component.hpp>
-#include <viam/sdk/log/logging.hpp>
 #include <viam/sdk/module/module.hpp>
 #include <viam/sdk/module/service.hpp>
 #include <viam/sdk/registry/registry.hpp>
 #include <viam/sdk/resource/resource.hpp>
 
 #include <third_party/trajectories/Trajectory.h>
+
+#include "utils.hpp"
 
 // this chunk of code uses the rust FFI to handle the spatialmath calculations to turn a UR vector to a pose
 extern "C" void* quaternion_from_axis_angle(double x, double y, double z, double theta);
@@ -107,23 +107,6 @@ void write_joint_data(const vector6d_t& jp, const vector6d_t& jv, std::ostream& 
     of << "\n";
 }
 
-// helper function to extract an attribute value from its key within a ResourceConfig
-template <class T>
-T find_config_attribute(const ResourceConfig& cfg, const std::string& attribute) {
-    std::ostringstream buffer;
-    auto key = cfg.attributes().find(attribute);
-    if (key == cfg.attributes().end()) {
-        buffer << "required attribute `" << attribute << "` not found in configuration";
-        throw std::invalid_argument(buffer.str());
-    }
-    const auto* const val = key->second.get<T>();
-    if (!val) {
-        buffer << "required non-empty attribute `" << attribute << " could not be decoded";
-        throw std::invalid_argument(buffer.str());
-    }
-    return *val;
-}
-
 // NOLINTNEXTLINE(performance-enum-size)
 enum class TrajectoryStatus { k_running = 1, k_cancelled = 2, k_stopped = 3 };
 
@@ -145,62 +128,6 @@ auto make_scope_guard(Callable&& cleanup) {
     return guard{std::forward<Callable>(cleanup)};
 }
 
-class URArmLogHandler : public urcl::LogHandler {
-   public:
-    URArmLogHandler() = default;
-    void log(const char* file, int line, urcl::LogLevel loglevel, const char* log) override {
-        std::ostringstream os;
-        os << "URCL - " << log_detail::trim_filename(file).data() << " " << line << ": " << log;
-        const std::string logMsg = os.str();
-
-        switch (loglevel) {
-            case urcl::LogLevel::INFO:
-                VIAM_SDK_LOG(info) << logMsg;
-                break;
-            case urcl::LogLevel::DEBUG:
-                VIAM_SDK_LOG(debug) << logMsg;
-                break;
-            case urcl::LogLevel::WARN:
-                VIAM_SDK_LOG(warn) << logMsg;
-                break;
-            case urcl::LogLevel::ERROR:
-                VIAM_SDK_LOG(error) << logMsg;
-                break;
-            case urcl::LogLevel::FATAL:
-                VIAM_SDK_LOG(error) << logMsg;
-                break;
-            default:
-                break;
-        }
-    }
-};
-
-void configure_logger(const ResourceConfig& cfg) {
-    std::string level_str{};
-    try {
-        level_str = find_config_attribute<std::string>(cfg, "log_level");
-    } catch (...) {
-        level_str = "debug";
-    }
-    const auto level = [&] {
-        if (level_str == "info") {
-            return urcl::LogLevel::INFO;
-        } else if (level_str == "debug") {
-            return urcl::LogLevel::DEBUG;
-        } else if (level_str == "warn") {
-            return urcl::LogLevel::WARN;
-        } else if (level_str == "error") {
-            return urcl::LogLevel::ERROR;
-        } else if (level_str == "fatal") {
-            return urcl::LogLevel::FATAL;
-        } else {
-            VIAM_SDK_LOG(error) << "invalid log_level: '" << level_str << "' - defaulting to 'debug'";
-            return urcl::LogLevel::DEBUG;
-        }
-    }();
-    urcl::setLogLevel(level);
-    urcl::registerLogHandler(std::make_unique<URArmLogHandler>());
-}
 }  // namespace
 
 void write_trajectory_to_file(const std::string& filepath, const std::vector<trajectory_sample_point>& samples) {
